统计图 支付优化

+++ src/main/java/com/neu/shop/controller/front/OrderController.java
@@ -39,6 +39,9 @@ public class OrderController {
     @Autowired
     private ActivityService activityService;
 
+    @Autowired
+    private CateService cateService;
+
     @RequestMapping("/order")
     public String showOrder(HttpSession session, Model model) {
 
@@ -68,7 +71,9 @@ public class OrderController {
 
         for (ShopCart cart:shopCart) {
             Goods goods = goodsService.selectById(cart.getGoodsid());
+            //获取商品库存
             int goodsNum = goods.getNum();
+            //判断库存是否充足
             if(goodsNum>=cart.getGoodsnum()){
                 List<ImagePath> imagePathList = goodsService.findImagePath(goods.getGoodsid());
                 goods.setImagePaths(imagePathList);
@@ -114,30 +119,36 @@ public class OrderController {
         ShopCartExample shopCartExample = new ShopCartExample();
         shopCartExample.or().andUseridEqualTo(user.getUserid());
         List<ShopCart> shopCart = shopCartService.selectByExample(shopCartExample);
-
-        //删除购物车
-        for (ShopCart cart : shopCart) {
-            shopCartService.deleteByKey(new ShopCartKey(cart.getUserid(),cart.getGoodsid()));
-        }
-
         //把订单信息写入数据库
-
         Order order = new Order(null, user.getUserid(),new Date(), oldPrice, newPrice, isPay, false, false, false, addressid,null,null);
         orderService.insertOrder(order);
         //插入的订单号
         Integer orderId = order.getOrderid();
 
-        //把订单项写入orderitem表中
         for (ShopCart cart : shopCart) {
-            orderService.insertOrderItem(new OrderItem(null, orderId, cart.getGoodsid(), cart.getGoodsnum()));
-        }
-        //库存减少
-        for (ShopCart cart : shopCart){
+            //获取购物车商品的预购数量
+            int payNum = cart.getGoodsnum();
+            //获取商品的库存
             Goods goods = goodsService.selectById(cart.getGoodsid());
             int goodsNum = goods.getNum();
-            int payNum = cart.getGoodsnum();
-            int goodsNewNum = goodsNum - payNum;
-            goodsService.updateGoodsNum(cart.getGoodsid(),goodsNewNum);
+            if(goodsNum>=payNum){
+                //把订单项写入orderitem表中
+                orderService.insertOrderItem(new OrderItem(null, orderId, cart.getGoodsid(), cart.getGoodsnum()));
+
+                //库存减少
+                int goodsNewNum = goodsNum - payNum;
+                int cateId = goods.getCategory();
+                goodsService.updateGoodsNum(cart.getGoodsid(),goodsNewNum);
+                //该类商品销售总量增加
+                Category category = cateService.selectById(cateId);
+                int saleNum = category.getSaleNum();
+                int newSaleNum = saleNum + payNum;
+                cateService.updateSaleNum(cateId,newSaleNum);
+                //删除相应购物车商品
+                shopCartService.deleteByKey(new ShopCartKey(cart.getUserid(),cart.getGoodsid()));
+            }else{
+                continue;
+            }
         }
         return Msg.success("购买成功");
     }
